requirements.txt
pandas
django-admin
django-crispy-forms
django-pandas
django-widget-tweaks
